// apps/enterprise-service-management/src/generated/models/field-validation-rule-model.ts
import type { FieldDefinition } from './field-definition-model';
import type { Workspace } from './workspace-model';

export const FieldValidationRuleRuleTypeKeyToLabel = {
  'MinimumLength': 'Minimum length',
  'MaximumLength': 'Maximum length',
  'MinimumValue': 'Minimum value',
  'MaximumValue': 'Maximum value',
  'Pattern': 'Pattern',
  'Email': 'Email',
  'Phone': 'Phone',
  'URL': 'URL'
} as const;
export type FieldValidationRuleRuleTypeKey = keyof typeof FieldValidationRuleRuleTypeKeyToLabel;

export const FieldValidationRuleStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type FieldValidationRuleStatusKey = keyof typeof FieldValidationRuleStatusKeyToLabel;

export interface FieldValidationRule {
  /**
   * @displayName Field Validation Rule ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Error Message
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  errorMessage: string;
  /**
   * @displayName Comparison Value
   * @validationRule Required for create/update operations
   */
  comparisonValue: string;
  /**
   * @displayName Created At
   */
  createdAt?: string;
  /**
   * @displayName Field Definition
   * @validationRule Required for create/update operations
   */
  fieldDefinition: Pick<FieldDefinition, 'id' | 'label'>;
  /**
   * @displayName Is Deleted
   */
  isDeleted?: boolean;
  /**
   * @displayName Rule Type
   * @validationRule Required for create/update operations
   */
  ruleTypeKey: FieldValidationRuleRuleTypeKey;
  /**
   * @displayName Sort Order
   */
  sortOrder?: number;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: FieldValidationRuleStatusKey;
  /**
   * @displayName Updated At
   */
  updatedAt?: string;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _FieldValidationRule = 'FieldValidationRule' as const;