// apps/enterprise-service-management/src/generated/models/form-section-model.ts
import type { FormDefinition } from './form-definition-model';
import type { FormVersion } from './form-version-model';
import type { Workspace } from './workspace-model';

export interface FormSection {
  /**
   * @displayName Form Section ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Title
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  title: string;
  /**
   * @displayName Created At
   */
  createdAt?: string;
  /**
   * @displayName Description
   * @validationRule Required for create/update operations
   */
  description: string;
  /**
   * @displayName Form Definition
   * @validationRule Required for create/update operations
   */
  formDefinition: Pick<FormDefinition, 'id' | 'name1'>;
  /**
   * @displayName Form Version ID
   * @validationRule Required for create/update operations
   */
  formVersionID: Pick<FormVersion, 'id' | 'snapshotLabel'>;
  /**
   * @displayName Instructions
   */
  instructions?: string;
  /**
   * @displayName Is Deleted
   */
  isDeleted?: boolean;
  /**
   * @displayName Sort Order
   * @validationRule Required for create/update operations
   */
  sortOrder: number;
  /**
   * @displayName Updated At
   */
  updatedAt?: string;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _FormSection = 'FormSection' as const;